fix(web): use component-level platform resolution for renderer sub-paths
