+++ src/gorhom/GorhomSheet.tsx
@@ -1,523 +1,4 @@
-import {
-  Fragment,
-  forwardRef,
-  useCallback,
-  useContext,
-  useEffect,
-  useId,
-  useImperativeHandle,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
-import { View, StyleSheet, useColorScheme, useWindowDimensions } from 'react-native';
+import type { WebRenderer } from '../web/types';
 
-import BottomSheet, {
-  BottomSheetBackdrop,
-  type BottomSheetBackdropProps,
-  BottomSheetFooter,
-  type BottomSheetFooterProps,
-  BottomSheetModal,
-  BottomSheetModalProvider,
-  type BottomSheetProps,
-  BottomSheetView,
-  type SNAP_POINT_TYPE,
-} from '@gorhom/bottom-sheet';
-import { useDerivedValue, useSharedValue } from 'react-native-reanimated';
-import type { ReactNode } from 'react';
-
-import { SheetContext } from '../web/SheetContext';
-import { getElevationShadow, renderSlot } from '../web/utils';
-import {
-  DEFAULT_CORNER_RADIUS,
-  DEFAULT_ELEVATION,
-  DEFAULT_MAX_WIDTH,
-  COLOR_SURFACE_CONTAINER_LOW_LIGHT,
-  COLOR_SURFACE_CONTAINER_LOW_DARK,
-  DEFAULT_ANCHOR_OFFSET,
-  DEFAULT_DETACHED_OFFSET,
-  DEFAULT_GRABBER_COLOR_LIGHT,
-  DEFAULT_GRABBER_COLOR_DARK,
-  DEFAULT_GRABBER_WIDTH,
-  DEFAULT_GRABBER_HEIGHT,
-} from '../web/constants';
-import type { TrueSheetRefMethods, WebRenderer } from '../web/types';
-import type {
-  TrueSheetProps,
-  DetentChangeEvent,
-  DidBlurEvent,
-  DidDismissEvent,
-  DidFocusEvent,
-  DidPresentEvent,
-  MountEvent,
-  PositionChangeEvent,
-  WillBlurEvent,
-  WillDismissEvent,
-  WillFocusEvent,
-  WillPresentEvent,
-  DragBeginEvent,
-  DragChangeEvent,
-  DragEndEvent,
-} from '../TrueSheet.types';
-
-const GorhomSheet = forwardRef<TrueSheetRefMethods, TrueSheetProps>((props, ref) => {
-  const {
-    name,
-    detents = [0.5, 1],
-    dismissible = true,
-    draggable = true,
-    dimmed = true,
-    dimmedDetentIndex = 0,
-    children,
-    scrollable = false,
-    initialDetentIndex = -1,
-    backgroundColor: backgroundColorProp,
-    cornerRadius = DEFAULT_CORNER_RADIUS,
-    elevation = DEFAULT_ELEVATION,
-    grabber = true,
-    grabberOptions,
-    maxContentHeight,
-    maxContentWidth,
-    anchor = 'center',
-    anchorOffset = DEFAULT_ANCHOR_OFFSET,
-    header,
-    headerStyle,
-    footer,
-    footerStyle,
-    onMount,
-    onWillPresent,
-    onDidPresent,
-    onWillDismiss,
-    onDidDismiss,
-    onDetentChange,
-    onPositionChange,
-    onDragBegin,
-    onDragChange,
-    onDragEnd,
-    onWillFocus,
-    onDidFocus,
-    onWillBlur,
-    onDidBlur,
-    detached,
-    detachedOffset = DEFAULT_DETACHED_OFFSET,
-    stackBehavior = 'switch',
-    style,
-  } = props;
-
-  const colorScheme = useColorScheme();
-  const backgroundColor =
-    backgroundColorProp ??
-    (colorScheme === 'dark' ? COLOR_SURFACE_CONTAINER_LOW_DARK : COLOR_SURFACE_CONTAINER_LOW_LIGHT);
-
-  const { width: windowWidth, height: windowHeight } = useWindowDimensions();
-  const isLandscapeOrTablet = windowWidth >= 600 || windowWidth > windowHeight;
-  const defaultName = useId();
-  const sheetName = name ?? defaultName;
-  const bottomSheetContext = useContext(SheetContext);
-  const bottomSheetModalRef = useRef<BottomSheetModal>(null);
-  const bottomSheetRef = useRef<BottomSheet>(null);
-  const initialDetentIndexRef = useRef(initialDetentIndex);
-  const currentIndexRef = useRef(0);
-  const isPresenting = useRef(false);
-  const isDismissing = useRef(false);
-  const isMinimized = useRef(false);
-  const isDragging = useRef(false);
-
-  const presentResolver = useRef<(() => void) | null>(null);
-  const dismissResolver = useRef<(() => void) | null>(null);
-
-  const animatedPosition = useSharedValue(windowHeight);
-  const animatedIndex = useSharedValue(0);
-
-  const [snapIndex, setSnapIndex] = useState(initialDetentIndex);
-  const [isMounted, setIsMounted] = useState(false);
-
-  const isNonModal = stackBehavior === 'none';
-
-  useDerivedValue(() => {
-    onPositionChange?.({
-      nativeEvent: {
-        position: animatedPosition.value,
-        index: animatedIndex.value,
-        detent: detents[animatedIndex.value] ?? 0,
-        realtime: true,
-      },
-    } as PositionChangeEvent);
-  });
-
-  const hasAutoDetent = detents.includes('auto');
-
-  const containerHeight = maxContentHeight ?? windowHeight;
-  const snapPoints = useMemo(
-    () =>
-      detents
-        .filter((detent): detent is number => detent !== 'auto' && typeof detent === 'number')
-        .map((detent) => Math.min(1, Math.max(0.1, detent)) * containerHeight),
-    [detents, containerHeight]
-  );
-
-  const handleChange = useCallback(
-    (index: number, _position: number, _type: SNAP_POINT_TYPE) => {
-      const previousIndex = currentIndexRef.current;
-      currentIndexRef.current = index;
-
-      // Handle drag end
-      if (isDragging.current && !isPresenting.current) {
-        isDragging.current = false;
-        onDragEnd?.({
-          nativeEvent: {
-            index,
-            position: animatedPosition.value,
-            detent: detents[index] ?? 0,
-          },
-        } as DragEndEvent);
-      }
-
-      if (!isPresenting.current && !isMinimized.current && previousIndex !== index && index >= 0) {
-        onDetentChange?.({
-          nativeEvent: {
-            index,
-            position: animatedPosition.value,
-            detent: detents[index] ?? 0,
-          },
-        } as DetentChangeEvent);
-      }
-
-      if (isPresenting.current) {
-        isPresenting.current = false;
-
-        // Resolve present promise
-        if (presentResolver.current) {
-          presentResolver.current();
-          presentResolver.current = null;
-        }
-
-        onDidPresent?.({
-          nativeEvent: {
-            index,
-            position: animatedPosition.value,
-            detent: detents[index] ?? 0,
-          },
-        } as DidPresentEvent);
-
-        onDidFocus?.({ nativeEvent: null } as DidFocusEvent);
-      }
-
-      // Fire onDidBlur when sheet reaches minimized state (index -1 but still mounted)
-      if (isMinimized.current && index === -1) {
-        onDidBlur?.({ nativeEvent: null } as DidBlurEvent);
-      }
-
-      // Fire onDidFocus when sheet is restored from minimized state
-      if (isMinimized.current && index >= 0) {
-        isMinimized.current = false;
-        onDidFocus?.({ nativeEvent: null } as DidFocusEvent);
-      }
-    },
-    [detents, animatedPosition]
-  );
-
-  const handleDismiss = useCallback(() => {
-    // Remove from stack when dismissed
-    bottomSheetContext?.removeFromStack(sheetName);
-
-    // Resolve dismiss promise
-    if (dismissResolver.current) {
-      dismissResolver.current();
-      dismissResolver.current = null;
-    }
-
-    onDidDismiss?.({ nativeEvent: null } as DidDismissEvent);
-
-    // Reset states since sheet is being dismissed
-    isMinimized.current = false;
-    isDismissing.current = false;
-    isDragging.current = false;
-  }, [sheetName]);
-
-  const handleAnimate = useCallback(
-    (_fromIndex: number, toIndex: number) => {
-      // Detect drag begin (when not presenting or dismissing)
-      if (!isPresenting.current && !isDismissing.current && !isDragging.current && toIndex >= 0) {
-        isDragging.current = true;
-        onDragBegin?.({
-          nativeEvent: {
-            index: currentIndexRef.current,
-            position: animatedPosition.value,
-            detent: detents[currentIndexRef.current] ?? 0,
-          },
-        } as DragBeginEvent);
-      }
-
-      // Drag change during animation
-      if (isDragging.current && toIndex >= 0) {
-        onDragChange?.({
-          nativeEvent: {
-            index: toIndex,
-            position: animatedPosition.value,
-            detent: detents[toIndex] ?? 0,
-          },
-        } as DragChangeEvent);
-      }
-
-      if (isPresenting.current) {
-        // Fire onMount on first present (before willPresent, matching native)
-        if (!isMounted) {
-          setIsMounted(true);
-          onMount?.({ nativeEvent: null } as MountEvent);
-        }
-
-        onWillPresent?.({
-          nativeEvent: {
-            index: toIndex,
-            position: animatedPosition.value,
-            detent: detents[toIndex] ?? 0,
-          },
-        } as WillPresentEvent);
-
-        // Focus events fire together with present events
-        onWillFocus?.({ nativeEvent: null } as WillFocusEvent);
-      }
-
-      // Detect if sheet is being restored (will focus)
-      if (isMinimized.current && toIndex >= 0) {
-        onWillFocus?.({ nativeEvent: null } as WillFocusEvent);
-      }
-
-      if (toIndex === -1 && !isPresenting.current) {
-        isMinimized.current = true;
-        onWillBlur?.({ nativeEvent: null } as WillBlurEvent);
-
-        // Only fire willDismiss if this is an actual dismiss (not being backgrounded by another sheet)
-        const sheetsAbove = bottomSheetContext?.getSheetsAbove(sheetName) ?? [];
-        if (sheetsAbove.length === 0) {
-          onWillDismiss?.({ nativeEvent: null } as WillDismissEvent);
-        }
-      }
-    },
-    [detents, animatedPosition, sheetName, bottomSheetContext, isMounted, onMount]
-  );
-
-  const backdropComponent = useCallback(
-    (backdropProps: BottomSheetBackdropProps) => {
-      if (!dimmed) {
-        return null;
-      }
-
-      // When not dismissible, collapse to below dimmed index instead of dismissing
-      const pressBehavior = dismissible
-        ? 'close'
-        : dimmedDetentIndex > 0
-          ? dimmedDetentIndex - 1
-          : 'none';
-
-      return (
-        <BottomSheetBackdrop
-          {...backdropProps}
-          opacity={0.5}
-          appearsOnIndex={dimmedDetentIndex}
-          disappearsOnIndex={dimmedDetentIndex - 1}
-          pressBehavior={pressBehavior}
-        />
-      );
-    },
-    [dimmed, dimmedDetentIndex, dismissible]
-  );
-
-  const indicatorHeight = grabberOptions?.height ?? DEFAULT_GRABBER_HEIGHT;
-
-  const handleStyle = useMemo(
-    () =>
-      grabber
-        ? [styles.handle, { paddingTop: grabberOptions?.topMargin }]
-        : { display: 'none' as const },
-    [grabber, grabberOptions?.topMargin]
-  );
-
-  const defaultGrabberColor =
-    colorScheme === 'dark' ? DEFAULT_GRABBER_COLOR_DARK : DEFAULT_GRABBER_COLOR_LIGHT;
-
-  const handleIndicatorStyle = useMemo(
-    () => ({
-      height: indicatorHeight,
-      borderRadius: grabberOptions?.cornerRadius ?? indicatorHeight / 2,
-      width: grabberOptions?.width ?? DEFAULT_GRABBER_WIDTH,
-      backgroundColor: grabberOptions?.color ?? defaultGrabberColor,
-    }),
-    [grabberOptions, indicatorHeight, defaultGrabberColor]
-  );
-
-  const footerComponent = useMemo(
-    () =>
-      footer
-        ? (footerProps: BottomSheetFooterProps) => (
-            <BottomSheetFooter
-              style={StyleSheet.flatten([styles.footer, footerStyle])}
-              {...footerProps}
-            >
-              {renderSlot(footer)}
-            </BottomSheetFooter>
-          )
-        : undefined,
-    [footer, footerStyle]
-  );
-
-  // For scrollable, we render the child directly
-  const ContainerComponent = scrollable ? Fragment : BottomSheetView;
-
-  const sheetMethodsRef = useRef<TrueSheetRefMethods>({
-    present: (index = 0) => {
-      return new Promise<void>((resolve) => {
-        presentResolver.current = resolve;
-        setSnapIndex(index);
-        isPresenting.current = true;
-        if (isNonModal) {
-          bottomSheetRef.current?.snapToIndex(index);
-        } else {
-          bottomSheetContext?.pushToStack(sheetName);
-          bottomSheetModalRef.current?.present();
-        }
-      });
-    },
-    dismiss: () => {
-      return new Promise<void>((resolve) => {
-        dismissResolver.current = resolve;
-        isDismissing.current = true;
-        if (isNonModal) {
-          bottomSheetRef.current?.close();
-        } else {
-          bottomSheetModalRef.current?.dismiss();
-        }
-      });
-    },
-    dismissStack: () => {
-      return new Promise<void>((resolve) => {
-        // Dismiss only sheets above, keeping this sheet presented
-        const sheetsAbove = bottomSheetContext?.getSheetsAbove(sheetName) ?? [];
-        const immediateChild = sheetsAbove[sheetsAbove.length - 1];
-        if (immediateChild) {
-          // Dismiss the immediate child - gorhom will dismiss all sheets above it
-          bottomSheetContext?.dismiss(immediateChild).then(resolve);
-          return;
-        }
-
-        resolve();
-      });
-    },
-    resize: async (index: number) => {
-      if (isNonModal) {
-        bottomSheetRef.current?.snapToIndex(index);
-      } else {
-        bottomSheetModalRef.current?.snapToIndex(index);
-      }
-    },
-  });
-
-  useImperativeHandle(ref, () => sheetMethodsRef.current);
-
-  // Register with context provider
-  useEffect(() => {
-    bottomSheetContext?.register(sheetName, sheetMethodsRef);
-    return () => {
-      bottomSheetContext?.unregister(sheetName);
-    };
-  }, [sheetName]);
-
-  // Auto-present on mount if initialDetentIndex is set
-  useEffect(() => {
-    if (initialDetentIndexRef.current >= 0) {
-      sheetMethodsRef.current.present(initialDetentIndexRef.current);
-    }
-  }, []);
-
-  const sheetContent = (
-    <ContainerComponent>
-      {header && <View style={headerStyle}>{renderSlot(header)}</View>}
-      {scrollable ? children : <View style={style}>{children}</View>}
-    </ContainerComponent>
-  );
-
-  const sharedProps: Omit<BottomSheetProps, 'children'> = {
-    style: [
-      styles.root,
-      {
-        backgroundColor,
-        borderTopLeftRadius: cornerRadius,
-        borderTopRightRadius: cornerRadius,
-        borderBottomLeftRadius: detached ? cornerRadius : 0,
-        borderBottomRightRadius: detached ? cornerRadius : 0,
-        boxShadow: getElevationShadow(elevation),
-        maxWidth: isLandscapeOrTablet ? (maxContentWidth ?? DEFAULT_MAX_WIDTH) : undefined,
-        marginLeft: isLandscapeOrTablet ? (anchor === 'left' ? anchorOffset : 'auto') : undefined,
-        marginRight: isLandscapeOrTablet ? (anchor === 'right' ? anchorOffset : 'auto') : undefined,
-        marginHorizontal: detached ? anchorOffset : undefined,
-      },
-    ],
-    backgroundComponent: null,
-    index: snapIndex,
-    enablePanDownToClose: dismissible,
-    enableContentPanningGesture: draggable,
-    enableHandlePanningGesture: draggable,
-    animatedPosition,
-    animatedIndex,
-    handleStyle,
-    handleIndicatorStyle,
-    onChange: handleChange,
-    onAnimate: handleAnimate,
-    enableDynamicSizing: hasAutoDetent,
-    maxDynamicContentSize: maxContentHeight,
-    snapPoints: snapPoints.length > 0 ? snapPoints : undefined,
-    detached,
-    bottomInset: detached ? detachedOffset : undefined,
-    backdropComponent,
-    footerComponent,
-  };
-
-  if (isNonModal) {
-    return (
-      <BottomSheet ref={bottomSheetRef} onClose={handleDismiss} {...sharedProps}>
-        {sheetContent}
-      </BottomSheet>
-    );
-  }
-
-  return (
-    <BottomSheetModal
-      ref={bottomSheetModalRef}
-      name={sheetName}
-      animateOnMount
-      onDismiss={handleDismiss}
-      stackBehavior={stackBehavior}
-      {...sharedProps}
-    >
-      {sheetContent}
-    </BottomSheetModal>
-  );
-});
-
-function GorhomProvider({ children }: { children: ReactNode }) {
-  return <BottomSheetModalProvider>{children}</BottomSheetModalProvider>;
-}
-
-export const GorhomRenderer: WebRenderer = {
-  Sheet: GorhomSheet,
-  Provider: GorhomProvider,
-};
-
-const styles = StyleSheet.create({
-  root: {
-    overflow: 'hidden',
-  },
-  handle: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    zIndex: 999,
-    paddingVertical: 10,
-    pointerEvents: 'none',
-  },
-  footer: {
-    pointerEvents: 'box-none',
-  },
-});
+// Gorhom renderer is web-only. On native, this is a no-op stub.
+export const GorhomRenderer = undefined as unknown as WebRenderer;
