+++ src/gorhom/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { GorhomRenderer } from './GorhomSheet';
