--- src/vaul/index.web.ts
@@ -1,1 +1,0 @@
-export { VaulRenderer } from './VaulSheet';
